src/common/requirements.txt
PyJWT==1.7.1
python-dotenv==0.19.2
mysql-connector==2.2.9
bcrypt==3.2.0
cffi==1.15.0
pycparser==2.21